uin16_t positition

+++ robolympics/cart_mega_code_i2c_wiringpi/cart_mega_code_i2c_wiringpi.ino
@@ -14,7 +14,7 @@ int test_case;
 int state;
 int ISR_state;
 int squarewave_pin = A7;
-int position;
+uint16_t position;
 
 bool print_data = false;
